Setup integration tests with Redis subprocess
taken from redis-rs.
Moved CLI and actual code apart (libbeachheadcompanion + beachhead-companion).
Initially intended as a way to test the core of the library independently, but that
didn't work out (test infrastructure would have to live in yet another crate).
The orphan rule was the last nail in the 'independent test infrastructure crate' coffin.